Fixed type of database

+++ database/schema.sql
@@ -45,7 +45,7 @@ CREATE TABLE Tickets (
 );
 
 CREATE VIEW AvailibleTickets AS (
-    SELECT e.eventid, (e.numtick - count(p.eventid)) as availibletickets FROM events e
+    SELECT e.eventid, (e.numtick - count(p.eventid))::int as availibletickets FROM events e
         INNER JOIN purchases p on e.eventid = p.eventid
         INNER JOIN tickets t on p.purchaseid = t.purchaseid
         GROUP BY e.eventid
